backend/app/requirements.txt
fastapi
uvicorn
httpx
pydantic
awkward
pymongo[srv]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
